add upload controller

+++ util/picManager.py
@@ -17,25 +17,33 @@ class picManager:
 
     def __init__(self, interval):
         self.blkBoard_mutex = threading.Lock()
-        self.blkBoard_listMutex = threading.Lock()
 
         self.pptPic_mutex = threading.Lock()
-        self.pptPic_listMutex = threading.Lock()
 
         self.interval = interval
 
+    def uploadPPT_pic(self,reqTime,image):
+        nowTime=str(int(time.time()))
+        #print "nowtime:"+nowTime
+        if reqTime+cc.VALID_PACK > int(nowTime):
+            if self.pptPic_mutex.acquire(cc.INTERVAL):
+                image.save(pc.PIC_LOCAL["2"]+nowTime+".jpeg")
+                self.pptPic_dict[nowTime]=nowTime
+                self.recentPic["2"]=nowTime
+                self.pptPic_mutex.release()
+            return 1
+        return 0
+
     def getBlkboard_pic(self, reqTime):
         image = b""
-        if abs(int(self.recentPic["blkBoard"]) - int(reqTime)) < self.interval:
-            picTime = self.recentPic["blkBoard"]
-
+        if reqTime + self.interval > int(self.recentPic["blkBoard"]):
             '''read file with file mutex
             '''
             if self.blkBoard_mutex.acquire(cc.INTERVAL):
+                picTime = self.recentPic["blkBoard"]
                 with open(pc.PIC_LOCAL["1"] + picTime + ".jpeg", "rb") as f:
                     image = f.read()
                 self.blkBoard_mutex.release()
-
         else:  # therer is no useful pic,then capture a new pic
             nowTime = str(int(time.time()))
             cap = cv.VideoCapture(cc.CAMERA_NUM)  # craete videoCapture object
@@ -45,32 +53,41 @@ class picManager:
                 '''
                 if self.blkBoard_mutex.acquire(cc.INTERVAL):
                     cv.imwrite(pc.PIC_LOCAL["1"] + nowTime + ".jpeg", frame)
+                    self.recentPic["blkBoard"] = nowTime
+                    self.blkPic_dict[nowTime] = nowTime  # add a new pic name to blkPic_list
                     self.blkBoard_mutex.release()
-
-                self.recentPic["blkBoard"] = nowTime
-            self.blkPic_dict[nowTime] = nowTime  # add a new pic name to blkPic_list
             cap.release()
             with open(pc.PIC_LOCAL["1"] + nowTime + ".jpeg", "rb") as f:
                 image = f.read()
         return image
 
     def getPPT_pic(self, reqTime):
-        pass
+        image=b""
+        if self.pptPic_mutex.acquire(cc.INTERVAL):
+            if int(self.recentPic["ppt"]) >0 and int(self.recentPic["ppt"])+pc.DELETE_INTERVAL> reqTime:
+                with open(pc.PIC_LOCAL["2"]+self.recentPic["ppt"]+"jpeg","rb") as f:
+                    image=f.read()
+            self.pptPic_mutex.release()
+        return image
 
     def DeleteUseless_pic(self):
-        for key in self.blkPic_dict.keys():
-            value = self.blkPic_dict[key]
-            nowTime = int(time.time())
-            if int(value) + pc.DELETE_INTERVAL < nowTime:
-                os.remove(pc.PIC_LOCAL["1"] + value + ".jpeg")
-                print "remove",value
-                self.blkPic_dict.pop(key)
-        for key in self.pptPic_dict.keys():
-            nowTime = int(time.time())
-            value = self.pptPic_dict[key]
-            if int(value) + pc.DELETE_INTERVAL < nowTime:
-                os.remove(pc.PIC_LOCAL["1"] + value + ".jpeg")
-                self.pptPic_dict.pop(key)
+        if self.blkBoard_mutex.acquire(cc.INTERVAL):
+            for key in self.blkPic_dict.keys():
+                value = self.blkPic_dict[key]
+                nowTime = int(time.time())
+                if int(value) + pc.DELETE_INTERVAL < nowTime:
+                    os.remove(pc.PIC_LOCAL["1"] + value + ".jpeg")
+                    print "remove",value
+                    self.blkPic_dict.pop(key)
+            self.blkBoard_mutex.release()
+        if self.pptPic_mutex.acquire(cc.INTERVAL):
+            for key in self.pptPic_dict.keys():
+                nowTime = int(time.time())
+                value = self.pptPic_dict[key]
+                if int(value) + pc.DELETE_INTERVAL < nowTime:
+                    os.remove(pc.PIC_LOCAL["1"] + value + ".jpeg")
+                    self.pptPic_dict.pop(key)
+            self.pptPic_mutex.release()
 
 '''
 regist pic type to picManager
